Adding Simple Framework for Creating Test ELFs (#29)
Created a framework for building RISC-V assembly tests for testing
Atlas. The framework uses magic memory to end simulation, which has not
been implemented in Atlas yet, so the `stop_sim_on_wfi` parameter is
required for these tests to end properly.

Created ELFs used in Atlas's regression should be checked in to the
`test/sim/workloads`dir.