First Stab at Supporting Sv32 and Sv39 Translation (#59)
This is very much a WIP. Much of this code was contributed by @skhanx86.
I've reworked some of his helper classes (`PageTableEntry` and
`PageTable`) and migrated some of his page walking code from the
`PageTableWalker` class into the translation unit. I added an Sv32
translation test that does a complete page table walk and translation of
a virtual address.