Implemented UART and Added Placeholder for Future Magic Memory Support (#76)
Added support for installing UART and Magic Memory devices in memory.
Provided parameters to enabling the UART and setting its base address
and size. Magic memory is enabled automatically if "tohost" and
"fromhost" symbols are found in the ELF.

I did not provide an implementation of Magic Memory in this PR. That
will be done later.