Supervisor Exceptions (#91)
Added support for taking exceptions in Supervisor mode. This is
controlled by the MEDELEG/SEDELEG CSRs for exceptions and the
MIDELEG/SIDELEG CSRs for interrupts. I also did a little bit of ground
work needed to support interrupts in the future. With this PR, the
"sbreak" arch test is now working correctly.