Remove outdated cosim comparison code (#140)
All of this code should be removed:
- SimDB has been rewritten and these header files no longer exist.
- This cosim comparison only supported the prototype Atlas IDE. No
remaining code reads these tables.
- The original cosim IDE was for Atlas/Imperas. We now have better
support in the new IDE for Atlas/Spike.

Rewriting this code just so that it compiles with the new SimDB doesn't
seem like a good use of time since nothing depends on it except a
deleted prototype UI.

This also unblocks Ayesha so she can use MAP latest instead of the ~5
week old MAP that still had the old SimDB.