Redesign the Action and ActionGroup Classes (#139)
To support multiple translation requests, the Translate Actions in an
instruction's ActionGroup need to be executed more than once. The
current ActionGroup design does not support this, because an executing
Action does not have a way to communicate that it needs to be executed
again. This PR reworks the ActionGroup "execute" method.

ActionGroup execution has been simplified. A ActionGroup contains a
vector of Actions that are executed in-order unless an ActionException
is thrown. When an ActionException is thrown, execution is interrupted
and redirected to a new ActionGroup. An ActionException will contain a
pointer to the next ActionGroup.