Program stack support + a few more syscalls/bug fixes (#165)
This work is moving along. It's still not running a Linux binary to the
end, but at least I'm getting to `main`!

I still don't have any testing for this as ELF programs are _still not
running_ to the end. Once I get that working,
then I'll be adding a LOT of testing.

`RunArchTests.py --riscv-arch` results:

```
    PASS RATE: 406/558
EXPECTED RATE: 406/558
```

---------

Co-authored-by: Knute Lingaard <klingaard@mips.com>