v ext: VectorState refactoring (#171)
This is the follow-up PR for initial VectorState refactoring.

- Now all vector groups / vector masks are `Elements`.
- `MaskElements` is a special `Elements`, in the sense that its
`Element` start and end position are calculated differently based on
`vstart` and `vl`.
- `MaskElementsIterator`, will adjust `Element` start and end position
based on `vstart` and `vl`. This is useful for vector mask bit-wise
operations.
- `MaskBitIterator` is used for iterating mask bit, when used as mask
for normal `Elements`.

Polymorphism is used. If vtable jump is a concern. We could switch to
template. But that approach has its own drawbacks.

---------

Co-authored-by: liya zhu <lzhu@mips.com>