CoSim pipeline (event retrieval only) (#199)
This PR starts the implementation of the CoSim flush / event retrieval
pipeline. Flushing will be added in the next PR, mostly needing to
change the `CoSimPipeline::createPipeline()` method and the
`cosim::Event` class to support it.

Note that this upstream PR needs to merge first:
https://github.com/sparcians/map/pull/594

Then MAP needs to cherry-pick and release the new SimDB (map_v2.1.0) so
Pegasus and others can point to it.

I'll keep this PR open for feedback in the meantime. The code shouldn't
have to change.