CMake Clean Up (#259)
Improved the way that the Pegasus libraries are organized so that we can
move to using `target_include_directories`. Added a new interface
library `pegasuslibs` to keep track of all of the included directories
needed for Pegasus to find header files. `pegasussim` is now the
"top-level" Pegasus library which links all of the Pegasus libraries
(e.g. pegasuscore, pegasussys) together with its external dependencies
(e.g. mavis, stf_lib).

Also moved the generated clang-format input file to the build directory.