Refactored Scripts to Generate Instruction JSONs and Header Files (#284)
To generate the Pegasus uarch JSONs, we were relying on the Python
instruction headers in `scripts/insts` lining up with the Mavis
instruction JSONs. This was brittle and required some hard-cording when
an extension had sub-extensions (e.g. "a" includes "zalrsc" and
"zaamo"). I cleaned this up so each of the Python headers now includes a
list of Mavis extensions support by this header.