Merge event pipeline and checkpointer for CoSim (#287)
The "pegasus_cosim_regress" target runs all ISA tests against the
cosim/flush harness. Current test results:
**PASS RATE: 86.67% (299/345)**

This is a couple weeks' worth of work so we don't have to wait for a
higher pass rate to get this checked in. I will work on the failing
tests and post smaller PRs for the fixes.