Implemented Pause Instruction and Added Multithreaded Test to Regression (#300)
Added support for the pause instruction, which stops the execution of a
thread until a pause counter expires. The counter's duration is in
cycles and can be set by a parameter. Also added a true multithreaded
test case to the regression, which requires a few miscellaneous fixes to
get working properly.

---------

Co-authored-by: Kathlene Magnus <kathlene.hurt@gmail.com>